requirements.txt
discord.py==1.3.3
praw==6.5.1     
prawcore==1.0.1
asyncio==3.4.3
psycopg2-binary==2.8.5
python-dotenv==0.13.0